requirements.txt
fastapi
uvicorn
python-dotenv
pybit